[EDGINREACH-48] - extracted test from InnReachProxyControllerTest to SecureTenantsProducerTest
